Quelques modifications sur position / déplacement de maladies

+++ VirusDefense/src/main/java/fr/iut/virusdefense/modele/maladie/BactérieBanale.java
@@ -4,7 +4,7 @@ import fr.iut.virusdefense.modele.Terrain;
 
 public class BactérieBanale extends Maladie{
 
-    public BactérieBanale(Terrain terrain, double x, double y){
+    public BactérieBanale(Terrain terrain, int x, int y){
         super(terrain, x, y, 10, 0.03);
     }
 }
